requirements.txt
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework-simplejwt
dotenv==0.9.9
pillow==11.3.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
gunicorn==21.2.0
python-decouple==3.8
dj-database-url==1.3.0
asgiref==3.10.0
pymysql==1.1.0
mysqlclient==2.2.0; sys_platform != 'win32'
psycopg2==2.9.10

